- move cli inside the package
