+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "tldrai"
-version = "0.2.0"
+version = "0.2.1"
 description = ""
 authors = ["omalinovskyi <orest0206@gmail.com>"]
 readme = "README.md"
@@ -17,7 +17,7 @@ pandas = "^2.2.2"
 lxml = "^5.2.2"
 
 [tool.poetry.scripts]
-tldrai = "cli:main_cli"
+tldrai = "tldrai.cli:main_cli"
 
 [tool.poetry.group.phi-1_5_Wizard_Vicuna_uncensored]
 optional = true
