+++ tests/test_search_functionality.py
@@ -1,5 +1,5 @@
 from click.testing import CliRunner
-from cli import main_cli
+from tldrai.cli import main_cli
 
 
 def test_cli_integration():
